Added FastNoise, TODO make it accept a long seed & improve terrain smoothing and test it out more

+++ src/main/java/com/azortis/orbis/generation/terrain/OIdTerrain.java
@@ -30,6 +30,6 @@ public interface OIdTerrain {
 
     String getName();
 
-    float getTerrainHeight(int x, int z, NoiseGenerator noiseGenerator);
+    double getTerrainHeight(int x, int z, NoiseGenerator noiseGenerator);
 
 }
